Added Technical Doc fix and title changes
Added fix to connect to technical document project

Added API Gallery and Interactive Chart titles
