+++ Projects/API Gallery/css/main.css
@@ -1,0 +1,162 @@
+/* Main CSS file for Project 10 */
+
+
+/* 320px breakpoint */
+
+#moveDataLabel {
+  display: inline-block;
+  width: 60%;
+  font-size: .7em;
+}
+#moveDataButton {
+  display: inline-block;
+  position: relative;
+  bottom: 17px;
+  width: 35%;
+  font-size: .5em;
+}
+#lightbox {
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  background-color: #756a75;
+  display: none;
+}
+.lightboxImg {
+  width: 75%;
+  height: 60%;
+  margin: 0 auto;
+  background-color: white;
+  color: #280C42;
+  position: relative;
+  top: 15%;
+  font-family: Arial, sans-serif;
+  font-weight: bolder;
+  line-height: 1.5;
+  overflow: auto;
+}
+#lightboxCloseWrapper {
+  width: 75%;
+  margin: 0 auto;
+  position: relative;
+  top: 15%;
+  color: #280C42;
+  font-size: 20px;
+}
+#lightboxClose {
+  display: inline-block;
+  position: relative;
+  top: 15%;
+  float: right;
+}
+.arrow {
+  border: solid black;
+  border-width: 0 3px 3px 0;
+  display: inline-block;
+  padding: 3px;
+  width: 20px;
+  height: 20px;
+}
+.right {
+  transform: rotate(-45deg);
+  -webkit-transform: rotate(-45deg);
+  position: relative;
+  bottom: 6rem;
+  left: 77%;
+}
+.left {
+  transform: rotate(135deg);
+  -webkit-transform: rotate(135deg);
+  position: relative;
+  bottom: 6rem;
+  left: 5%;
+}
+#lightboxCaption {
+  position: relative;
+  width: calc(75% - 1px);
+  top: calc(11% - 1px);
+  left: calc(12% - -4px);
+  background-color: white;
+  color: #280C42;
+  font-family: Tahoma, sans-serif;
+  line-height: 1.2;
+}
+#header {
+  width: 80%;
+  margin: 0 10%;
+  padding: 20px 0;
+  font-size: 30px;
+  border-bottom: solid 1px;
+}
+.dataAll {
+  display: inline-block;
+  overflow: auto;
+  width: 92%;
+  height: 20%;
+  background: #fff;
+  color: #280C42;
+  border: solid 1px;
+  margin-top: 10px;
+  margin-left: -4%;
+  margin-bottom: 10px;
+  font-family: Arial, sans-serif;
+  font-weight: bolder;
+  line-height: 1.5;
+}
+
+/* 768px breakpoint */
+
+@media screen and (min-width: 768px) {
+  #moveDataButton {
+    bottom: 4px;
+    left: 21px;
+  }
+  #lightboxCloseWrapper, .lightboxImg, #lightboxCaption {
+    width: 80%;
+  }
+  #lightboxCaption {
+    left: 10%;
+    line-height: 1.5;
+  }
+  .left {
+    left: 4%;
+  }
+  .right {
+    left: 88%;
+  }
+  .dataAll {
+    margin-left: 5%;
+    width: 84%;
+  }
+}
+
+/* 1024px breakpoint */
+
+@media screen and (min-width: 1024px) {
+  #header {
+    width: 90%;
+    margin: 0 5%;
+    margin-bottom: 10px;
+  }
+  #wrapper {
+    width: 90%;
+    margin: 0 5%;
+  }
+  right {
+    left: 90%;
+  }
+  .dataAll {
+    overflow: scroll;
+    height: 250px;
+  }
+  .dataAll:nth-child(odd) {
+    margin-top: 0;
+    margin-left: calc(-4% - 4px);
+    width: 49%;
+  }
+  .dataAll:nth-child(even) {
+    margin-top: 0;
+    margin-left: 10px;
+    width: 49%;
+  }
+}
