+++ Projects/API Gallery/index.html
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>API Gallery</title>
+  <meta content="width=device-width, initial-scale=1.0" name="viewport">
+  <link href="css/app.min.css" rel="stylesheet" type="text/css"> 
+  <script src="https://code.jquery.com/jquery-3.2.1.min.js" type="text/javascript">
+  </script>
+  <script src="https://code.jquery.com/ui/1.12.0/jquery-ui.min.js" type="text/javascript">
+  </script>
+  <script src="js/app.min.js" type="text/javascript">
+  </script>
+</head>
+<body>
+  <div id="lightbox">
+    <div id="lightboxCloseWrapper">
+      Press X to close
+      <div id="lightboxClose">
+        X
+      </div>
+    </div>
+    <div class="lightboxImg"></div>
+    <div class="left arrow" id="leftArrow"></div>
+    <div class="right arrow" id="rightArrow"></div>
+    <div id="lightboxCaption">
+      Title:
+    </div>
+  </div>
+  <div id="header">
+    My Favorite Data on the Web<br>
+    <label id="moveDataLabel">Want to see Pokemon data first?</label> <button id="moveDataButton">Gotta see them all!</button>
+  </div>
+  <ol id="wrapper">
+    <li class="dataAll" id="swData1" value="0"></li>
+    <li class="dataAll" id="swData2" value="1"></li>
+    <li class="dataAll" id="swData3" value="2"></li>
+    <li class="dataAll" id="swData4" value="3"></li>
+    <li class="dataAll" id="swData5" value="4"></li>
+    <li class="dataAll" id="swData6" value="5"></li>
+    <li class="dataAll" id="swData7" value="6"></li>
+    <li class="dataAll" id="swData8" value="7"></li>
+    <li class="dataAll" id="pokeData1" value="8"></li>
+    <li class="dataAll" id="pokeData2" value="9"></li>
+    <li class="dataAll" id="pokeData3" value="10"></li>
+    <li class="dataAll" id="pokeData4" value="11"></li>
+    <li class="dataAll" id="pokeData5" value="12"></li>
+    <li class="dataAll" id="pokeData6" value="13"></li>
+    <li class="dataAll" id="pokeData7" value="14"></li>
+    <li class="dataAll" id="pokeData8" value="15"></li>
+  </ol>
+</body>
+</html>
