+++ projects.html
@@ -75,7 +75,7 @@
           <!-- API Gallery Project-10 -->
           <div class="projects-card-layout">
             <div class="card" style="width: 18rem;">
-              <a href="./Projects/Project%2010/index.html" target="_blank"><img class="card-img-top" src="./img/Project-10.jpg" alt="Project Image"></a>
+              <a href="./Projects/API%20Gallery/index.html" target="_blank"><img class="card-img-top" src="./img/Project-10.jpg" alt="Project Image"></a>
               <div class="card-body">
                 <h5 class="card-title">API gallery</h5>
                 <p class="card-text">Added different API's with a lightbox using HTML, CSS, and JavaScript</p>
@@ -85,7 +85,7 @@
           <!-- Chart.JS Project-9 -->
           <div class="projects-card-layout">
             <div class="card" style="width: 18rem;">
-              <a href="./Projects/Project%209/index.html" target="_blank"><img class="card-img-top" src="./img/Project-9.jpg" alt="Project Image"></a>
+              <a href="./Projects/Interactive%20Chart.js/index.html" target="_blank"><img class="card-img-top" src="./img/Project-9.jpg" alt="Project Image"></a>
               <div class="card-body">
                 <h5 class="card-title">Interactive Chart.js front page</h5>
                 <p class="card-text">Developed an interactive page using Chart.js using HTML, CSS, and JavaScript</p>
@@ -153,7 +153,7 @@
           <!-- Technical Document -->
           <div class="projects-card-layout">
             <div class="card" style="width: 18rem;">
-              <a href="./Projects/TechDoCPage/index.html" target="_blank"><img class="card-img-top" src="./img/techDocPage.png" alt="Project Image"></a>
+              <a href="./Projects/TechDocPage/index.html" target="_blank"><img class="card-img-top" src="./img/techDocPage.png" alt="Project Image"></a>
               <div class="card-body">
                 <h5 class="card-title">Technical Document</h5>
                 <p class="card-text">Created a sleak Technical Document Page with fixed sidebar and links to each section using HTML and CSS</p>
